♻️ 5H7BAA code: split context ingest task helpers
